Add --fmt-only flag
This is a hacky one-off implementation to help Twitter deal with the
fact that they have a custom Goal called scalafix, which does the
equivalent of `fmt --fmt-only=scalafix`, as it provides them with a
forward-compatible way of migrating people off of their scalafix goal.

When the v2 goal is renamed from fmt2 to fmt, this option should be
moved to that Goal, and its implementation broadened to support all v2
formatters, as well as any v1 formatters we fancy while we keep them
around.