Rename `optionable_rule` to `subsystem_rule` (#8893)
### Rationale

This rename is meant to make rule authoring more familiar for Pants rule authors. 

`Subsystem`s are a concept used frequently in both V1 and V2 which most plugin authors are familiar with. It is also a concrete concept. In V2, developers exclusively use `GoalSubsystem`s and `Subsystem`s to consume options, so the name `subsystem` is familiar.

In contrast, `Optionable` is an interface that most developers never use, unless they go under-the-hood to see how subsystems work. Likewise, `Optionable` may be easily confused for `Optional`.

### Possible future need for flexibility

There could possibly be a future where we have some new type of `Optionable` that is not a `Subsystem` or `GoalSubsystem`. Were that to happen, then `subsystem_rule` name would not work because it's not as flexible as the current `optionable_rule`.

Rather than this being a reason to stick with the status quo, however, we could instead introduce in this future a new rule name that explicitly describes that concept, including possibly reusing `optionable_rule` to co-exist with `subsystem_rule`. 

Until this possible future, only having `subsystem_rule` makes the concept more concrete and leverages the prior knowledge of Pants devs.