Bump up deprecations for the 1.26 series to earlier in the 1.26 dev series (#8908)
_**This has no impact on stable releases.**_

Our deprecation policy never required two full cycles for deprecations, but, due to a misunderstanding, we were doing this in some cases. For example, if we deprecated something in 1.0.0.dev2, we were waiting until 1.2.0.dev2, rather than 1.2.0.dev0, to remove the code.

This misunderstanding was slowing down deprecations necessary for incrementally migrating from V1 to V2. Here, we slightly bump up those deprecations planned for the `1.26.0.dev` series to happen earlier in that series.

We also update the docs to better clarify how the policy works.