Rename `Spec` to `AddressSpec` (#8916)
Prework for https://github.com/pantsbuild/pants/issues/8593 to support non-target specs, e.g. file system specs.

## Rationale
We don’t want to expand the current definition of `Spec` to include files - it’s a useful abstraction as is, e.g. that we have logic for `SingleAddress`.

At the same time, “specs” is a useful namespace to describe your _specifications_ for what you want to run against. We want to be able to use `Specs` to describe an abstraction `specification` and then have `AddressSpec` and `FilesystemSpec` describe the concrete specifications, e.g.:

![hierarchy tree for Spec](https://user-images.githubusercontent.com/14852634/71929478-f2e0f900-3156-11ea-92dd-50c9e8d41f28.png)

(Note that the design for `FilesystemSpec` will likely change - this is the first draft.)

### Rejected alternative: `PositionalArg`
`PositionalArg` is an implementation detail - the real fundamental idea of this object is that this is a specification for something to run against.

## Re: deprecation policy
This is safe to change without a deprecation cycle because `Spec` is not marked as a public API and its only direct usage is in rules, which are experimental.

We do not rename functions that are public to reflect this name change. Namely, we keep `Target.compute_dependency_specs()` as the same name rather than `Target.compute_dependency_address_specs()`.