A rule to topologically sort targets.
Useful in some limited cases, but in general target
dependency order is enforced via rule graph dependency order,
so most rules don't need to consider dep order directly.

One use case for this is in the setup-py rule - we need to
be able to order exported targets so we publish dependencies
before dependees.