Use color in v2 pytest (#8901)
## Problem

When we run python tests in V2, we run a pex that invokes pytest as a python library to actually do the testing, using the ExecuteProcessRequest abstraction. When run this way, pytest detects that it's not being run in a terminal and so doesn't normally display color.

## Solution

Add the environment variable PYTEST_ADDOPTS, setting it to --color=yes if the global color option is set, or to --color=no if it's not. This forces (or unforces) pytest to output color regardless of the state of the terminal.

## Result

This results in V2 python testing displaying color output in the same way it did in V1.