Update core test model to include coverage results (#8851)
### Problem

To get coverage reports working I need to be able to access the coverage data file dropped in the current working directory when tests are run.

### Solution

Add a field to `TestResult` to allow accessing `process_result.output_directory_digest`. For now, we mark this as a private Python-specific field. In the future, we will decide what this field should look like.

### Result

We can get coverage reports with the V2 python test runner. (separate PR incoming.)