Simplify `./pants test --debug` requesting one single address (#8925)
Because the rule to get one single `BuildFileAddress` took `AddressSpecs` as its subject, we needed a little extra ceremony in `rules/core/test.py` to get a single `BuildFileAddress`. 

This was unnecessary, though. The rule `single_build_file_address` only ever needs `BuildFileAddresses` and never the additional information from `AddressSpecs`. Making the change to the rule signature allows us to thus simply pass `BuildFileAddresses` as the subject in `test.py`.