Deprecate `--positional-arg-file` in favor of `--spec-file` (#8928)
In preparation for non-target specs, we had renamed `--target-spec-file` to `--positional-arg-file`. 

However, in https://github.com/pantsbuild/pants/pull/8916, we decided to use the name `Spec` to describe a general _specification_ for what Pants should run against, rather than `PositionalArg`. This updates the option and all relevant code to use the newly preferred name `spec`.

We also explain in the deprecation warnings why this change is happening so that users don't think we're introducing all this churn for no good reason.