Have handle_workunits callback use kwargs (#8918)
## Problem

We want the streaming workunits abstraction to be able to signal that the last batch of workunits is going to be sent, which requires a change to the signature of the handle_workunits callback that we expect Subsystems to provide. More generally we want any further changes to the contract that StreamingWorkunitsHandler makes with the callback to not affect existing Subsystems that make use of that functionality.

## Solution

Change the streaming workunit subsystem to always call the handle_workunits callback with a set of kwargs, and document exactly what those kwargs are.

As an additional change, modify the scope of the streaming workunits session() context, so that the try-finally block that shuts down run_tracker is within the scope. This makes sure that the last workunits message sent from the streaming workunit handler happens after RunTracker shuts down.

## Result

This will require a one-time change in any client Subsystems making use of the streaming workunits functionality, but afterwards pants will be able to add additional kwargs to this callback without requiring any change to client Subsystems.