Share the local LMDB store to avoid creating one per test method. (#8933)
### Problem

Running enough unit tests that extend `TestBase` will eventually exhaust the virtual memory* on your machine and report a "No space left on device" error on OSX (*it's unconfirmed that this is _precisely_ what is happening, but there are no "real" files taking up enough space to cause the issue).

### Solution

Do not create anonymous local store directories for every scheduler we initialize in unit tests. While this was a good idea from a "it's better to have more hermeticity if it doesn't hurt" perspective, it does appear to hurt.

The `file` and `directory` store databases (see `~/.cache/pants/lmdb_store/`) are 100% harmless to share at this level, since they are content addressed, and not the subject of python-level unit tests. The `processes` store database is more questionable: but I'd argue that this is the wrong level to be directly testing that cache at anyway.

### Result

No more "No space left on device" errors.