Add `create_options_bootstrapper()` test util for less boilerplate in V2 tests (#8943)
V2 tests that depend on `subsystem_rules` suffer from substantial boilerplate, requiring these things to work properly:

1) Every subsystem used transitively must be registered as a `RootRule`.
2) Every subsystem used transitively must be initialized via `testutil.subsystem.util.init_subsystem`.
3) Every subsystem used transitively must be passed as a `Param` to the rule.

Instead, we can pass an `OptionsBootstrapper` as a `Param` so that the engine works properly with the `subsystem_rules` and avoids all three of these steps. This is how `ConsoleRuleTestBase` does things already.