Fix _maybe_handle_help (#8945)
## Problem

We were previously calling _maybe_handle_help within the with context provided by the streaming workunit handler. _maybe_handle_help checks to see if the user is trying to invoke ./pants goals and if so prints top-level goal help and calls _exiter, which calls sys.exit. This means that we were exiting the program within the with-context when the user ran ./pants goals, which meant that the with-context wasn't doing its job of shutting down the thread that polled for workunit information. So, if the user ran ./pants goals the program wouldn't cleanly exit but the workunit polling thread would stick around looping forever.

## Solution

Have _maybe_handle_help return a signal value rather than directly calling the exit function, and check for that signal value outside the with-context to determine if we even need to enter it. Also refactor some of the code that was in the with-context to make debugging a bit easier in the future.

Co-authored-by: Asher Foa <asher@toolchain.com>