Skip hanging Pantsd integration test (#8961)
This started hanging consistently around 11 AM PST on the master branch and several PR builds. 

The first failure was for https://travis-ci.org/pantsbuild/pants/builds/636504715, which only adds f-strings and is highly unlikely to be causing the hang. 

A PR to run each test method as an individual target isolated the issue to `test_sigint_kills_request_waiting_for_lock()`. We skip this for now to stabilize CI.