Use Python 3.6+ by default for Python users (#8968)
Users of the Python backend will now default to Python 3.6+ for their interpreter constraints, which impacts for example which interpreter is used for `./pants test`. Before, we defaulted to Python 2.7 _or_ Python 3.6+.

To go back to the original, users should set `['CPython>=2.7,<3', 'CPython>=3.6']` under `python-setup` in their `pants.ini`.

This also updates the default Pytest to use Pytest 5, which only works with Python 3 tests.