Tweak options parser to allow globs in filesystem specs (#8981)
### Problem
Originally, we stated that we do _not_ want to add support for globs in filesystem specs because it would be too difficult to support. Instead, users should rely on their shell's builtin glob expansion.

However, `PathGlobs`—which is how we'll implement filesystem specs—has built-in support for globs. So, it is valid for a user to specify `./pants cloc2 'src/python/**/*.py'` (note the quotes around the spec). In fact, it would be difficult for us to restrict users from using `./pants cloc2 'src/python/**/*.py'` because we would need to proactively validate that they aren't using globs, even though the engine does support them.

### Solution
Our `ArgSplitter` heuristic already supported globs, outside of the top-level `*` spec. This fixes that edge case and adds testing to ensure globs work at the `ArgSplitter` level.