Wire up ParsedSpecs instead of TargetRoots
`cmd_line_spec_parser.py` parses both address specs and filesystem specs regardless of V1 vs. V2. This commit adds support for that parsing and wires up `ParsedSpecs` so that it's provided in all the right places.

At this point, we are not actually consuming the filesystem specs, only ensuring that they are available.

This requires changing the `cmd_line_spec_parser.py` heuristics:
* If the spec has an extension, assume it's a file. Address specs should never have extensions in their name.
* If there is `*`, assume it's a glob.
* If there is ambiguity between the shorthand `a/b -> a/b:b` vs. the file `a/b`, then use `Path.is_file()` to resolve if it's a real file or not.

We _do not_ attempt to change the semantics of what a directory mean, e.g. `a/b` where `a/b` is a real directory (not a file).