Remove the generate-pants-ini goal.
It can be replaced by a shell one-liner.

The reasoning behind removing it is that it is only ever
run at most once ever per-repo, by whoever sets up the repo
for the first time. After that it's just dead weight, but since
it appears in every `./pants goals` listing, it's just confusing
for all users for all time. We should optimize for a good
ongoing user experience, and this goal just adds to their
cognitive load ("here's a thing you don't need to care about")
without providing any significant benefit to that initial set up.
After all, editing pants.ini manually is something codebase admins
are going to have to do sooner or later anyway.