Allow for deprecation of child scopes
### Problem

Currently an attempt to deprecate a scope below another scope (eg: deprecating the `test.mytask` scope in favor of `mytask`) will fail 50% of the time based on registration order due to the requirement that scopes are registered in alphabetical order, and governed by a requirement that parent scopes are frozen before children.

These requirements mean that if the new scope for a deprecation sorts before the old scope, the deprecated sub scope will already have options registered in it by the time the parent is visited, triggering a `FrozenRegistration` error.

### Solution

This form of deprecation is very useful (the "`test`" example above is real-ish, and triggered by welcome v2 changes in the `test` scope). The parser freezing case appeared to be in service of an optimization, but removing it and replacing it with eager lookups to possibly shadowed options 1) is clearer, 2) has equivalent performance.

### Result

Nested option scopes may be deprecated. Fixes #8984.