Set workunit parent IDs correctly (fixes #7969) (#8947)
## Problem

Cf. #7969 , we weren't previously setting a workunit's parent ID correctly.

## Solution

The solution turned out to be to make sure that we call workunit_store::[get|set]_parent_id in task_executor, to make sure that the task-local notion of what the most current parent ID for a workunit is gets propagated to newly-spawned tasks. It also makes sure that at least one workunit is the first one to start and set the parent id, which ensures that that workunit itself will have a None parent_id in Python.