Fix Zipp 2.0 release breaking Python 2 integration tests (#8994)
Zipp recently released 2.0, which drops support for Python 2. Zipp is a dependency of `importlib_metadata`, which is a dependency of `pytest`. This meant that our Pytest integration tests which run with Python 2 were failing to resolve the PEX.

This issue should be fixed by Pex 2.0's improved dependency resolution, but, in the meantime, this stabilizes CI.