Restore support for `dependencies` outputting 3rdparty deps through new enum option (#8960)
Turns out, there is a use case for `./pants dependencies --external-only`, such as scripts using it to determine the exact version strings of Python requirements used by Pants targets.

However, the option names are not ideal, such as having to enforce that they're mutually exclusive at runtime. It's also a bit confusing to default to `internal and external`. Ideally, we'd default to `sources` (aka `internal`) and only include `3rdparty` (aka `external`) when requested.

We can make both improvements through a new enum option, which wasn't available when we first wrote this due to Python 2 not having enums.

![help message for the dependencies2 goal](https://user-images.githubusercontent.com/14852634/72855006-54cf5180-3c73-11ea-9351-29d7e880384a.png)