Change distribution lookup for JvmTask subclasses to call into JvmTask (#8938)
### Problem
A number of tasks ought to use JVM versions based on what a target requires via it's kwargs. Compile and test tasks both explicitly use target declared JVM platforms, but there are a number of other tasks that should depend on them as well. Compile is a special case and wasn't modified here.

I'm working on making it possible to declare runtime platforms separately from compile platforms so that you can, for example, migrate the runtime version separately from the compile version of the JDK.

### Solution

This patch consolidates the platform lookups for JvmTask subclasses so that platform lookups all happen through one set of code paths. This makes it a lot simpler to introduce a new concept of a `runtime_platform` to JVM targets.

Specifically it does the following
    - Change all DistributionLocator usages in JvmTask subclasses with calls into preferred_jvm_distribution_for_targets
    - Add a cut out so that junit_run can use test_platform
    - Pass JDK through to JvmPlatform calls so that javadoc can be covered, though javadoc invokes won't work with version of the JDK > 8 (#8937)

### Result

JvmTasks use a JVM that matches the platform declared on targets, and all of them use a common code path to look up those platforms.