Include the line number
This allows jumping to line, rather than having to grep the function name.

However, this naively would cause an issue with tests. Were we to hardcode the line number, the tests would become exceptionally brittle. Any change to the file would risk causing most tests to fail.

To address this, we introduce the helper functions fmt_task_func() and fmt_rule(). It's not great how tightly coupled this is with the engine, but it reduces the surface area of that coupling to only these two new methods.

(Wish I knew to make these helper methods from the start...)