s/ParsedSpecs/Specs
Stu had a point that ParsedSpecs is unnecessarily wordy. Of course they're parsed from the raw strings!

I was a little hesitant to use the name `Specs` because I was originally thinking that we would also need to implement `Spec`. This is not true. `Specs` should be a dataclass that wraps `address_specs` and `filesystem_specs`, not a `Collection[Spec]`.