Refactor `config.py` to allow supporting possible future config file formats (#9014)
We already were going out of our way to encapsulate our config abstraction and decouple from `configparser`. Our `Config` API is very similar to `configparser`, but encapsulates the library and provides a limited API rather than allowing direct access to `configparser`.

The only place this encapsulation broke through was in `_SingleFileConfig`, where `options_bootstrapper.py` was directly calling methods on a `configparser.ConfigParser` instance, rather than relying on an interface provided by Pants.

This PR extends the prior encapsulation so that `config.py` is the only file ever interacting with the `configparser` module. Beyond better encapsulation, this gives us the possibility to add support for other config formats like TOML in the future.