Deprecate `--fast` option (#8970)
This change is meant to give a nudge to users in the direction of the V2 test runner. By having them switch to `--no-fast` now, we lower the barrier to entry to the V2 test runner, which must run with `--no-fast`. 

This is meant to avoid a Python 2 vs. Python 3 situation where the cost to migrate from the V1 test runner to the V2 implementation is too costly such that users stick with V1 and the Pants community now bifurcates.