Modify options parser to allow `!` in file args (#9031)
We will soon allow `!` in both the `sources` field of BUILD files and in file args (aka filesystem specs) to indicate excludes. This PR tweaks our options parsing logic to formalize support for `!` as a leading character.

This also lightly refactors `engine/legacy/structs.py` and some `PathGlobs`-related code in preparation for adding support to the `!` syntax.