Removing the exception when setup_py task has no work to do. (#8620)
Problem:
The setup_py task raises a TaskError when it does not have any tasks to do. This is causing a problem when I am developing a DockerBuild plugin that will consume the python wheels that were created by the SetupPy task.

My DockerBuild task marks the output of SetupPy as optional_data input. However not all docker image targets I am generating actually consume a generated python wheel. In such cases, the SetupPy is still invoked by the RoundManager which then fails with a TaskError.

Solution:
To fix this issue, I modified the SetupPy task to not raise a TaskError when there is nothing to do.

Result:
As a result, the SetupPy task behaves like all other tasks that do not have any targets it can action on and quietly returns.