Refactor `PathGlobs` to support `!` ignore globs in its constructor (#9032)
### Problem
Per https://github.com/pantsbuild/pants/issues/5427, we are generally unifying the way you mark includes and excludes into a single flattened list of globs, where the leading `!` in `!ignore.py` is used to indicate excludes.

It was proposed in https://github.com/pantsbuild/pants/pull/9025#pullrequestreview-349626588 that we extend this unification to `PathGlobs` as well. 

### Solution
Change the Python `PathGlobs` dataclass to have a `globs` field, rather than `include` and `exclude` fields.

Notably, we do not change the actual `PathGlobs` Rust struct. There, it is helpful to still differentiate between includes and excludes. Instead, this changes the Python API and `PathGlobs::create`.

### Result
It should be easier to add support for `!` ignores to BUILD files now. For no extra work, we now support `!` ignores with filesystem specs.

This change also allows rule authors to use the same style of ignores we now use in BUILD files and in filesystem specs when working with `PathGlobs`.