No longer use `globs`, `rglobs`, and `zglobs` internally (#9029)
Per https://github.com/pantsbuild/pants/issues/5427, we will be deprecating `globs`, `rglobs` and `zglobs` in BUILD files. This PR prepares the repo for that change.

Most importantly, this automates the process through a new script `build-support/migration-support/fix_deprecated_globs_usage.py`. This task is highly tedious but also highly automatable. Thanks to the script, we were able to update all 104 of our BUILD files in less than 3 minutes (2 minutes of manual changes flagged by the script).

When we deprecate `globs`, we will include instructions to curl the script and run it so that our users don't have to waste time updating BUILD files. 

Because we plan to distribute the `fix_deprecated_globs_usage.py` (and because it's complex), we add exhaustive unit tests.