Deprecate `--glob-expansion-failure` in favor of `--files-not-found-behavior` (#9022)
As proposed in https://docs.google.com/document/d/15xphZcFnowytF0Qu2sO1PG7wHAWoyLof7hK0jVa5T8E/edit?disco=AAAAI4KUmss, we decided to remove the ability for users to ignore glob failures. It is not safe for them to ignore failures as it often leads to subtle bugs.
We've recently improved glob match errors—such as describing the origin of the failure—such that we can now be comfortable requiring users to deal with unmatched globs in their BUILD files.

Notably, we still allow rule authors to use `GlobMatchErrorBehavior.ignore`. This only impacts the options we expose to end users.

--

This also changes file args (aka filesystem specs) to always warn no matter what, rather than using whatever `--files-not-found-behavior` is configured to. Those should not be coupled because they are very different use cases.