Stop using `globs` in testprojects/ (#9039)
Originally, we didn't update `testprojects/` to stop using `globs` et al. because some of the testprojects intentionally use `globs`. 

But, turns out, there are about 200 deprecated usages of `globs` (mostly from the chroot test migration) which may be safely updated.

This also makes the `fix_deprecated_globs_usage.py` script more robust. It was failing when a BUILD file had functions declared in it and with the argument `excludes=set(['foo.py'])`.