Snapshot subset intrinsic (#9006)
## Problem

This is prework for #8956 . We want to be able to request a Digest that contains a subset of the files and directories from another Snapshot.

## Solution

Create a new root type SnapshotSubset, which wraps a Digest and a list of files/dirs from that Digest to be put into a new Digest. Then create an engine intrinsic that will filter stored file information to that subset and output a new Digest.