[IntelliJ] Add compiler options and plugins to the output of `export-dep-as-jar` (#9036)
### Problem
In order to faithfully reproduce the compilation of targets, tools that consume export-dep-as-jar need access to the compiler options and compiler plugins used by pants.

Context on compiler options:
Compiler options (for scalac) are defined in pants.ini under scope compile.rsc, in what we call "compiler option sets".
A target may choose to enable or disable some of these option sets. Therefore, compiler options are per-target.
Context on how pants implements using scalac plugins: here

### Solution
Consume the zinc_args product in export-dep-as-jar. (credit to @stuhood that pointed me in the direction of prior art)
Filter all the zinc_args to leave only the scalac ones (the ones starting with -S), and strip the prefix to leave pure scalac args.
Output it, per target, in export-dep-as-jar.

### Result
The entries for modulizable targets in the output of export-dep-as-jar now contain a field, compiler_options, that contains the aggregated command line information that would be passed to scalac.

Example output:

❯ ./pants --pants-config-files='+["examples/src/scala/org/pantsbuild/example/scalac/compiler_options/pants.compiler_options.ini"]' export-dep-as-jar examples/src/scala/org/pantsbuild/example/scalac/compiler_options:with_nonfatal_warnings_and_plugin

...
{
    "version": "1.0.14",
    "targets": {
        "examples/src/scala/org/pantsbuild/example/scalac/compiler_options:with_nonfatal_warnings_and_plugin": {
            "id": "examples.src.scala.org.pantsbuild.example.scalac.compiler_options.with_nonfatal_warnings_and_plugin",
            ...
            "compiler_options": [
                "-Ywarn-unused",
                "-Xplugin:.pants.d/compile/rsc/c1e3836b60e5/examples.src.scala.org.pantsbuild.example.scalac.plugin.simple_scalac_plugin/current/zinc/classes:.pants.d/compile/rsc/c1e3836b60e5/examples.src.scala.org.pantsbuild.example.scalac.plugin.other_simple_scalac_plugin/current/zinc/classes:.pants.d/compile/rsc/c1e3836b60e5/examples.src.scala.org.pantsbuild.example.scalac.plugin.simple_scalac_plugin_helper/current/zinc/classes",
                "-P:simple_scalac_plugin:args",
                "-P:simple_scalac_plugin:from",
                "-P:simple_scalac_plugin:target",
                "-P:simple_scalac_plugin:local"
            ],
            ...
        }
    },
    ...
}