Decouple `--owner-of` implementation from` --changed-*` (#9046)
### Problem

Soon, we will be deprecating `--owner-of` in favor of filesystem specs. To do this, we will likely end up using `OwnersRequest` or a derivative in the rule to go from `FilesystemSpecs -> BuildFileAddresses`.

However, currently, the rule `find_owners(OwnersRequest) -> BuildFileAddresses` is much more complex than necessary because it is able to calculate both `--owner-of` and `--changed-*`, where `--changed-*` is a more complex calculation. As a result, the rule `find_owners()` has direct dependencies on `AddressMapper` and `BuildConfiguration`, along with several `await Get`s never used by `--owner-of`.

We want `--owner-of` to be as simple as possible so that it may be more easily leveraged for filesystem specs. 

### Solution

Simplify `OwnersRequest` to no longer have a field `include_dependees` and to only ever find direct owners. Add a new `changed.find_owners()` rule which will use the `OwnersOf` rule to compute the direct owners, then add its own logic to calculate dependencies and combine the results.

This builds heavily off of https://github.com/pantsbuild/pants/pull/8917 but is not as ambitious due to concerns with caching of `SCM`: https://github.com/pantsbuild/pants/pull/8917/files#r364023925

### Result

`--owner-of` and `--changed-*` behave identically to before, but have less tight of coupling. `--changed-*` still depends on `--owner-of`, but `--owner-of` now no longer has any dependency on `--change-*`.

This should make filesystem specs support easier to add.