Deprecate `--build-file-imports` defaulting to `warn` and having the option `allow` (#9047)
### Problem
Imports are not safe in BUILD files, as described in:

https://github.com/pantsbuild/pants/pull/9040#discussion_r372910342

We did not have a mechanism to warn about them until December 2017 via https://github.com/pantsbuild/pants/pull/5180, so we seem to have only maintained the misfeature due to backward compatibility. 

### Solution
We want to remove support for them, but most do so incrementally per the deprecation policy. First, we deprecate `--build-file-imports=allow` and at the same time deprecate the default being `warn` instead of `error`. Then, in 1.27.0.dev0, we will change the default to `error` and deprecate the option entirely.

### Result
We will have remove a whole class of bugs for users. We'll also be able to make more assumptions about BUILD files. From #5180:

> It also opens up potential future opimisations such as parsing with a
custom parser rather than a whole python environment.