Remove origin check from release.sh script (#9042)
## Problem

The release.sh script runs a check to see if the "origin" remote is one of several variants of the URL for the official Github pants repo. However, nothing else in the script makes any reference to a remote locally-called "origin" - the function that actually creates the tag (rightly) refers to the offiical Github repo with a fully-qualified URL. This means that the release.sh script breaks for anyone who names their reference to the official repo with something other than "origin" (I use "upstream" for this, for instance).
## Solution

Simply remove this check.