Fix V2 `test` and `setup-py2` to work with file args (#9056)
### Problem

`test.py` and `run_setup_py.py` both use the type `AddressProvenanceMap`, which itself uses `ProvenancedBuildFileAddresses`. There is no way to go from `FilesystemSpecs -> ProvenancedBuildFileAddresses`—only `FilesystemSpecs -> BuildFileAddresses`—so file args don't work with the `test` and `setup-py2` goals.

Generally, `ProvenancedBuildFileAddress` is a way to preserve the "provenance" (or origin) of how a `BuildFileAddress` was resolved:

https://github.com/pantsbuild/pants/blob/1ebe376f1a4d672c9955e7b80c8a6caba48a7719/src/python/pants/engine/addressable.py#L20-L24

When resolving `AddressSpecs`, we roughly follow this flow: `AddressSpecs -> ProvenancedBuildFileAddresses -> BuildFileAddresses`, unlike `FilesystemSpecs -> BuildFileAddresses`.

### Solution

Tweak `ProvenancedBuildFileAddress`'s definition to use `Spec` rather than `AddressSpec`, so that the origin may come from a `FilesystemSpec`.

Then, add a rule to go from `FilesystemSpecs -> ProvenancedBuildFileAddresses`. It would be very difficult to backtrace the original `FilesystemSpec` that resulted in the owning `BuildFileAddress` (https://github.com/pantsbuild/pants/issues/9055), so for now we provide a no-op implementation. Currently, this `provenance` would never be consumed for Filesystem Specs so this is safe to do.

### Result

_All_ goals now work with file args.

If we decide in https://github.com/pantsbuild/pants/issues/9055 that we do care about preserving the provenance for filesystem specs enough to re-architect how filesystem args are implemented, then we have a foundation to build off of.